backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai>=1.0.0
pillow==10.3.0
pillow-heif==0.18.0
pdf2image==1.17.0
openpyxl==3.1.2
pydantic==2.5.3
